refactor: move built in commands and components to handlers package
